test: move pad_to_size CHW tests into augment_test.py
Address review feedback on PR #112:
- Move the channels-first pad_to_size regression tests out of the
  standalone pad_to_size_test.py and into the existing TestCustom class
  in augment_test.py, matching that file's conventions.
- Drop the explanatory comments, consistent with the surrounding tests.

Delete dm_pix/_src/pad_to_size_test.py.
